Implemented 13.2
Implemented defocus blur for camera. Instead of firing rays from an exact point, rays are now fired randomly from a disk at camera position, hitting the viewport inaccurately both before and past the lookat point. Both the size of the disk and the distance between the camera and the viewport can be modified as parameters to change focus behavior.